refactor: merge parser and ui test builder functionality together and update imports
